Remove SecurityCheckActivity & run checks in SecureActivity
